+++ web_search_neo/__init__.py
@@ -1,3 +1,3 @@
 """Web Search Neo package."""
 
-__version__ = '1.18.0'
+__version__ = '1.18.1'